styling the NavBar component
