--- src/components/Header/index.js
@@ -1,15 +1,0 @@
-import React from "react";
-import { Link } from "react-router-dom";
-import CurrencySelect from "../CurrencySelect";
-
-const Header = () => {
-  return (
-    <>
-      <main>
-        <CurrencySelect />
-      </main>
-    </>
-  );
-};
-
-export default Header;
