+++ src/components/NavBar/NavBar.styles.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+import { Row, Col } from "antd";
+
+export const Nav = styled.nav``;
+
+export const Container = styled(Row)``;
